backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database connectors
psycopg2-binary==2.9.10
mysql-connector-python==9.1.0
sqlalchemy==2.0.36

# LLM providers
google-genai>=1.0.0
openai>=1.0.0
anthropic>=0.40.0

# Encryption
cryptography==43.0.3

# Validation / settings
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# SQL parsing
sqlglot==25.33.0

# Rate limiting
slowapi>=0.1.9

# Async support
anyio==4.7.0

# Testing
pytest==8.3.4
httpx==0.28.1
